+++ src/main/java/org/betonquest/betonquest/BetonQuest.java
@@ -238,6 +238,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.Event;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -1526,6 +1527,51 @@ public class BetonQuest extends JavaPlugin {
         } catch (final InstructionParseException e) {
             log.warn(pack, "&cCould not create variable '" + name + "': " + e.getMessage(), e);
             return "";
+        }
+    }
+
+    /**
+     * Resolves the global variable using the global data context.
+     * Supports both globaltag and globalpoint variable types.
+     *
+     * @param packName name of the package
+     * @param name     name of the variable (without % characters, including type as a prefix)
+     * @return the value of this global variable
+     */
+    @SuppressWarnings("PMD.AvoidLiteralsInIfCondition")
+    public @NotNull String getGlobalVariableValue(final @Nullable String packName, final @NotNull String name) {
+        if (!Config.getPackages().containsKey(packName)) {
+            log.warn("Global variable '" + name + "' contains the non-existent package '" + packName + "' !");
+            return "";
+        }
+        final String[] nameParts = name.split("\\.");
+        if (nameParts.length != 2) {
+            log.warn("Global variable '" + name + "' does not specify type! Use a '.' as name delimiter!");
+            return "";
+        }
+        switch (nameParts[0]) {
+            case "globaltags":
+            case "globaltag":
+            case "gtag":
+            case "gtags":
+            case "gt":
+                final String tag = packName + "." + nameParts[1];
+                return Boolean.toString(globalData.getTags().contains(tag));
+            case "globalpoints":
+            case "globalpoint":
+            case "gpoints":
+            case "gpoint":
+            case "gp":
+                final String category = packName + "." + nameParts[1];
+                for (final Point point : globalData.getPoints()) {
+                    if (point.getCategory().equals(category)) {
+                        return Integer.toString(point.getCount());
+                    }
+                }
+                return "";
+            default:
+                log.warn("Global variable '" + name + "' type is invalid!");
+                return "";
         }
     }